Fix trailing whitespace trimming in files with mixed line endings
When a file with mixed line endings is opened in Vim 'filetype' is set
to 'unix' and CR characters are shown as part of the text. This breaks
trimming of trailing whitespace on lines ending with CRLF because now CR
is the last character on such lines and it's not considered a whitespace
character.
To work around that we should consider a sequence of whitespace
characters followed by a single CR at the end of line as trailing
whitespace (excluding CR character itself).